a18nadapter renamed to a18nclient
