--- src/A18NAdapter/Dto/EmailMessageList.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace A18NAdapter.Dto
-{
-    public class EmailMessageList
-    {
-        public List<EmailMessage> EmailMessages { get; set; }
-        public int Count { get; set; }
-    }
-}
