--- src/A18NAdapter/Dto/SmsMessageList.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace A18NAdapter.Dto
-{
-    public class SmsMessageList
-    {
-        public List<SmsMessage> SmsMessages { get; set; }
-        public int Count { get; set; }
-    }
-}
