+++ src/A18NClient/A18NClient.csproj
