+++ src/A18NClient/Dto/EmailMessage.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace A18NAClient.Dto
+{
+    public class EmailMessage
+    {
+        public string MessageId { get; set; }
+        public string ProfileId { get; set; }
+        public string ToAddress { get; set; }
+        public string FromAddress { get; set; }
+        public DateTime CreatedAt { get; set; }
+        public string Subject { get; set; }
+        public string Content { get; set; }
+    }
+}
