+++ src/A18NClient/Dto/EmailMessageList.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace A18NAClient.Dto
+{
+    public class EmailMessageList
+    {
+        public List<EmailMessage> EmailMessages { get; set; }
+        public int Count { get; set; }
+    }
+}
