+++ src/A18NClient/Dto/SmsMessage.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace A18NAClient.Dto
+{
+    public class SmsMessage
+    {
+        public string MessageId { get; set; }
+        public string ProfileId { get; set; }
+        public string Sender { get; set; }
+        public string Receiver { get; set; }
+        public DateTime CreatedAt { get; set; }
+        public string Content { get; set; }
+        public string Url { get; set; }
+    }
+}
