+++ src/Okta.Idx.Sdk.E2ETests/Okta.Idx.Sdk.E2ETests.csproj
@@ -28,7 +28,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\A18NAdapter\A18NAdapter.csproj" />
+    <ProjectReference Include="..\A18NClient\A18NClient.csproj" />
   </ItemGroup>
 
   <ItemGroup>
